CS251_SSL/Project/Float-Moodle/floatmoodle/requirements.txt
absl-py==0.13.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.3.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
imutils==0.5.4
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mediapipe==0.8.6
mistune==0.8.4
monotonic==1.5
MouseInfo==0.1.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.4.0
numpy==1.17.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.54
packaging==20.9
pafy==0.5.5
pandas==1.3.2
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAutoGUI==0.9.52
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyDrive==1.3.1
PyGetWindow==0.0.9
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyRFC3339==1.1
pyrsistent==0.17.3
PyScreeze==0.1.27
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
scipy==1.7.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.4.2
stevedore==3.3.0
systemd-python==234
terminado==0.9.5
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
youtube-dl==2021.6.6
